Implement API root endpoint
API root endpoint returns list of available API versions and
current version.

Closes: ansible/galaxy-dev#80
