+++ tests/api/test_api_views.py
@@ -1,0 +1,11 @@
+from .base import BaseTestCase, API_PREFIX
+
+
+class TestApiRootView(BaseTestCase):
+    def test_get(self):
+        response = self.client.get(f"/{API_PREFIX}/")
+
+        assert response.status_code == 200
+        assert response.data == {
+            "available_versions": {"v3": f"/{API_PREFIX}/v3/"},
+        }
